--- a/be/src/exprs/runtime_filter.cpp
+++ b/be/src/exprs/runtime_filter.cpp
@@ -1145,7 +1145,7 @@ Status IRuntimeFilter::get_prepared_vexprs(std::vector<vectorized::VExpr*>* vexp
            (_state->enable_pipeline_exec() &&
             _rf_state_atomic.load(std::memory_order_acquire) == RuntimeFilterState::READY));
     DCHECK(is_consumer());
-    std::lock_guard<std::mutex> guard(_inner_mutex);
+    std::lock_guard guard(_inner_mutex);
 
     if (_push_down_vexprs.empty()) {
         RETURN_IF_ERROR(_wrapper->get_push_vexprs(&_push_down_vexprs, _state, _vprobe_ctx));
@@ -1179,7 +1179,7 @@ bool IRuntimeFilter::await() {
         }
     } else {
         SCOPED_TIMER(_await_time_cost);
-        std::unique_lock<std::mutex> lock(_inner_mutex);
+        std::unique_lock lock(_inner_mutex);
         if (_rf_state != RuntimeFilterState::READY) {
             int64_t ms_since_registration = MonotonicMillis() - registration_time_;
             int64_t ms_remaining = wait_times_ms - ms_since_registration;
@@ -1187,8 +1187,18 @@ bool IRuntimeFilter::await() {
             if (ms_remaining <= 0) {
                 return false;
             }
+#if !defined(USE_BTHREAD_SCANNER)
             return _inner_cv.wait_for(lock, std::chrono::milliseconds(ms_remaining),
                                       [this] { return _rf_state == RuntimeFilterState::READY; });
+#else
+            auto timeout_ms = butil::milliseconds_from_now(ms_remaining);
+            while (_rf_state != RuntimeFilterState::READY) {
+                if (_inner_cv.wait_until(lock, timeout_ms) != 0) {
+                    // timeout
+                    return _rf_state == RuntimeFilterState::READY;
+                }
+            }
+#endif
         }
     }
     return true;
@@ -1242,7 +1252,7 @@ void IRuntimeFilter::signal() {
     if (_state->enable_pipeline_exec()) {
         _rf_state_atomic.store(RuntimeFilterState::READY);
     } else {
-        std::unique_lock<std::mutex> lock(_inner_mutex);
+        std::unique_lock lock(_inner_mutex);
         _rf_state = RuntimeFilterState::READY;
         _inner_cv.notify_all();
     }