--- a/be/src/io/cache/whole_file_cache.cpp
+++ b/be/src/io/cache/whole_file_cache.cpp
@@ -36,7 +36,8 @@ WholeFileCache::WholeFileCache(const Path& cache_dir, int64_t alive_time_sec,
 
 WholeFileCache::~WholeFileCache() {}
 
-Status WholeFileCache::read_at(size_t offset, Slice result, const IOContext& io_ctx, size_t* bytes_read) {
+Status WholeFileCache::read_at(size_t offset, Slice result, const IOContext& io_ctx,
+                               size_t* bytes_read) {
     if (bthread_self() == 0) {
         return read_at_impl(offset, result, io_ctx, bytes_read);
     }
@@ -47,7 +48,7 @@ Status WholeFileCache::read_at(size_t offset, Slice result, const IOContext& io_
 }
 
 Status WholeFileCache::read_at_impl(size_t offset, Slice result, const IOContext& io_ctx,
-                               size_t* bytes_read) {
+                                    size_t* bytes_read) {
     if (io_ctx.reader_type != READER_QUERY) {
         return _remote_file_reader->read_at(offset, result, io_ctx, bytes_read);
     }