--- a/be/src/io/fs/file_system_map.h
+++ b/be/src/io/fs/file_system_map.h
@@ -1,0 +1,49 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+#pragma once
+
+#include <memory>
+#include <shared_mutex>
+#include <unordered_map>
+
+#include "io/fs/file_system.h"
+#include "util/lock.h"
+
+namespace doris {
+namespace io {
+
+class FileSystemMap {
+public:
+    static FileSystemMap* instance();
+    ~FileSystemMap() = default;
+
+    void insert(ResourceId id, FileSystemSPtr fs);
+
+    // If `id` is not in `_map`, return nullptr.
+    FileSystemSPtr get(const ResourceId& id);
+
+private:
+    FileSystemMap() = default;
+
+private:
+    doris::SharedMutex _mu;
+    std::unordered_map<ResourceId, FileSystemSPtr> _map; // GUARED_BY(_mu)
+};
+
+} // namespace io
+} // namespace doris