--- a/be/src/io/fs/local_file_reader.cpp
+++ b/be/src/io/fs/local_file_reader.cpp
@@ -20,6 +20,7 @@
 #include <atomic>
 
 #include "util/doris_metrics.h"
+#include "util/async_io.h"
 
 namespace doris {
 namespace io {
@@ -39,7 +40,13 @@ Status LocalFileReader::close() {
     bool expected = false;
     if (_closed.compare_exchange_strong(expected, true, std::memory_order_acq_rel)) {
         DorisMetrics::instance()->local_file_open_reading->increment(-1);
-        auto res = ::close(_fd);
+        int res = -1;
+        if (bthread_self() == 0) {
+            res = ::close(_fd);
+        } else {
+            auto task = [&] { res = ::close(_fd); };
+            AsyncIO::run_task(task, io::FileSystemType::LOCAL);
+        }
         if (-1 == res) {
             return Status::IOError("failed to close {}: {}", _path.native(), std::strerror(errno));
         }
@@ -47,9 +54,19 @@ Status LocalFileReader::close() {
     }
     return Status::OK();
 }
-
-Status LocalFileReader::read_at(size_t offset, Slice result, const IOContext& /*io_ctx*/,
+Status LocalFileReader::read_at(size_t offset, Slice result, const IOContext& io_ctx,
                                 size_t* bytes_read) {
+    if (bthread_self() == 0) {
+        return read_at_impl(offset, result, io_ctx, bytes_read);
+    }
+    Status s;
+    auto task = [&] { s = read_at_impl(offset, result, io_ctx, bytes_read); };
+    AsyncIO::run_task(task, io::FileSystemType::LOCAL);
+    return s;
+}
+
+Status LocalFileReader::read_at_impl(size_t offset, Slice result, const IOContext& /*io_ctx*/,
+                                     size_t* bytes_read) {
     DCHECK(!closed());
     if (offset > _file_size) {
         return Status::IOError("offset exceeds file size(offset: {}, file size: {}, path: {})",