--- a/be/src/olap/lru_cache.cpp
+++ b/be/src/olap/lru_cache.cpp
@@ -196,7 +196,7 @@ void LRUCache::_lru_append(LRUHandle* list, LRUHandle* e) {
 }
 
 Cache::Handle* LRUCache::lookup(const CacheKey& key, uint32_t hash) {
-    std::lock_guard<std::mutex> l(_mutex);
+    std::lock_guard l(_mutex);
     ++_lookup_count;
     LRUHandle* e = _table.lookup(key, hash);
     if (e != nullptr) {
@@ -219,7 +219,7 @@ void LRUCache::release(Cache::Handle* handle) {
     LRUHandle* e = reinterpret_cast<LRUHandle*>(handle);
     bool last_ref = false;
     {
-        std::lock_guard<std::mutex> l(_mutex);
+        std::lock_guard l(_mutex);
         last_ref = _unref(e);
         if (last_ref) {
             _usage -= e->total_size;
@@ -359,7 +359,7 @@ Cache::Handle* LRUCache::insert(const CacheKey& key, uint32_t hash, void* value,
     THREAD_MEM_TRACKER_TRANSFER_TO(e->total_size, tracker);
     LRUHandle* to_remove_head = nullptr;
     {
-        std::lock_guard<std::mutex> l(_mutex);
+        std::lock_guard l(_mutex);
 
         // Free the space following strict LRU policy until enough space
         // is freed or the lru list is empty
@@ -402,7 +402,7 @@ void LRUCache::erase(const CacheKey& key, uint32_t hash) {
     LRUHandle* e = nullptr;
     bool last_ref = false;
     {
-        std::lock_guard<std::mutex> l(_mutex);
+        std::lock_guard l(_mutex);
         e = _table.remove(key, hash);
         if (e != nullptr) {
             last_ref = _unref(e);
@@ -425,7 +425,7 @@ void LRUCache::erase(const CacheKey& key, uint32_t hash) {
 int64_t LRUCache::prune() {
     LRUHandle* to_remove_head = nullptr;
     {
-        std::lock_guard<std::mutex> l(_mutex);
+        std::lock_guard l(_mutex);
         while (_lru_normal.next != &_lru_normal) {
             LRUHandle* old = _lru_normal.next;
             _evict_one_entry(old);
@@ -452,7 +452,7 @@ int64_t LRUCache::prune() {
 int64_t LRUCache::prune_if(CacheValuePredicate pred) {
     LRUHandle* to_remove_head = nullptr;
     {
-        std::lock_guard<std::mutex> l(_mutex);
+        std::lock_guard l(_mutex);
         LRUHandle* p = _lru_normal.next;
         while (p != &_lru_normal) {
             LRUHandle* next = p->next;