@hello-stephen 2023-01-17
TeamCity pipeline, clickbench performance test result:
 the sum of best hot time: 36.52 seconds
 load time: 491 seconds
 storage size: 17121910321 Bytes
 https://doris-community-test-1308700295.cos.ap-hongkong.myqcloud.com/tmp/20230117201530_clickbench_pr_82584.html